Add Gateway config
The gateway config is responsible for interfacing with the Magento
codebase to retrieve the settings for the komoju plugin and present them
through an API. Updating the di.xml to insert the payment method code
(komoju_payments) into the Config programmatically, as it is a constant
and needs to be kept the same everwhere through the code.

Updating the ConfigProvider to use the new Config class to retrieve the
config data, as opposed to the Braintree Config, which was a copy/paste
error. Since the ConfigProvider will be specifically used to inject the
config into the frontend for the UI rendering, the secret key and
webhook secret token have been removed, as exposed those in the browser
would have been a major security flaw.