Render payment type options based on plugin config
Only display Credit Card and Konbini options as possible payment types
if they are enabled in the plugin settings. Currently if both are
disabled it will just render an empty select box, so additional logic
will be applied later to not render Komoju as a payment option at all if
neither Credit Card or Konbini are selected.