Clean up config
Removed some (for now) unnecessary config around validators, and finally
solved the mystery of why my Config wasn't working. Because I was still
passing in the base Magento config class in the di.xml, which is why it
didn't work until I updated the config to pass the method code into the
base Config class. This was due to a copy paste error, which is a good
reminder to be more careful when I'm copy code from the example plugins.