Add Redirect action to send user to the Hosted Page
Adding a controller action to create the Hosted Page API url and
redirect the user to it. This is being navigated to after the customer
has clicked submit and the order data has been saved by Magento, so
we're retrieving the Order and Billing details, getting the required
fields from them and then generating the HostedPage URL. Magento follows
a similar pattern for matching controllers to routes that Rails does, so
this file can be navigated to at <site-url>/komoju/hostedpage/redirect.

This is just an initial PoC of getting the redirect to work, so the
return and cancel URLs still need to be populated with the correct data,
and the external order num needs to be updated to be more specific to
Magento.