Ignore additional plugins for git
Adding the ngrok plugin to the gitignore list, since it's an external
plugin not related to this one. I intend to go back and add installing
it as a step in the dev Dockerfile, but for now it's just easier to app
it to the src/app/code directory locally.