Add Cancel route
Adding a route to cancel the order when the cancel button on the hosted
page is clicked. Because the order has already been created in Magento
by the time the user is redirected to the hosted page if the order is
not explicitly cancelled when they choose not to enter their payment
details it will just stay in the pending state permanently.