Configure the HostedPage url with correct redirects
When the user cancels the payment, redirect to the cancel route, which
will cancel the order in the system and then redirect to the store front
page. When the payment succeeds the user is redirected to the success
page, which is the default success page for all checkout methods. It
doesn't look like Magento exposes that as a constant which is why it's
hard coded, but Magento has hardcoded the checkout/onepage/success route
multiple places in their own codebase, so this seems like an acceptable
choice.