Generate redirect url dynamically
Previously the redirect url on the JS frontend was hardcoded, but this
won't work as it needs to be specific to it's current magento store.
Generating the URL in the config and inserting into the
window.checkoutConfig to be used by JS allows the URL to be created for
the current instance.