Add HMAC validation to the Cancel endpoint
Because the cancel endpoint takes an order Id and cancels it in the
system it's important that we place some safety checks on it's use,
otherwise unsavoury characters could use it to cause problems for the
Magento shop owners. By using the Komoju Secret Key and generating a
HMAC param when the cancel URL is created in the Redirect controller we
can ensure that someone cannot just pass in any order Id and cancel an
order.