Cancel order only when in correct states
Only allow the cancellation of an order when the order is in the
pending_payment state (which is the default state a Komoju payments
order is in until getting webhook events) or cancelled. This will
prevent the endpoint from being used to cancel complete orders or put
the order into an impossible state.