Change allowed state from pending to new
Turns out that the current implementation creates an order with a state
of new, instead of pending_payment as I originally thought. This meant
that the code would not cancel legitimate orders when the cancel button
was clicked on the hosted page. Updating the check to use the new state
const instead.