Add model for Webhook events
Adding a model file to encapsulate the webhook event JSON data into a
consistent interface that can be used throughout the codebase, so if the
JSON request structure changes then we just need to update the model to
reflect the changes and not have to change all uses of the model within
the codebase.

This is pretty much a straight copy from the version in the
komoju-woocommerce plugin, with some small changes to be consistent with
this plugin's styling.