Update order status when payment captured
When the app receives a payment.captured event from Komoju then the
payment is complete, and the order is marked in the system as complete.
This is the initial event to verify that the process works, and will be
expanded to cover additional event types and fleshed out with more
functionality in later commits.