Change the captured state to processing
When looking at the entire order lifeline, an order wouldn't be
considered closed until the order had shipped, so marking the order
complete once payment has been received is a bit premature. Instead,
it's being marked as processing which in the magento lifecycle means the
payment has been successful and the order is ready to be fulfilled.