Mark order as processing before redirect
Before redirecting the user to the hosted page we're now updating the
order to be in a pending payment state, since this more accurately
reflects the expected order status flow for payments in Magento.
Updating the Cancel route logic to consider pending payment to be valid
instead of new payments, to ensure that the cancellation still works as
expected.