Change order status based on event type
Update the order status according to the event type. When the event has
been cancelled/expired/failed then change the order state to cancelled,
as there's no additional steps the store admins can take. If the event
is unknown then return a 400, so we can track if the plugin needs to
accept additional events easier, rather than have it just ignore the
errors.