Add comments on payment state change
Adding order history comments when processing events. Because all the
data in the webhooks are available in the Komoju dashboard we don't have
to copy it all into the order status, but providing messages to let the
admins know when and why an order has changed state is helpful for them
to know what happened without having to go the Komoju dashboard.