Refactor webhook processing into separate file
Refactoring the webhook processing into it's own code. The Plugin/
directory appears to be where any code that doesn't fit into any other
folder (like a lib/), which feels like the best place for this code,
since it doesn't fit cleanly into any of the MVC definitions. As part of
the refactor a custom exception was created, UnknownEventException. When
an object that extends the default Magento exception is thrown to the
top of the call stack then Magento just returns it to the requester,
using the defined status code. This saves on having to manually catch an
exception in the controller and craft an non 200 status code to return.