Add table to hold external payments id
Currently the external order num passed to Komoju when redirecting to
the Hosted Page url is the order id of the order, which is a
monotonically increasing number. The problem with this solution is that
multiple storefronts could have that same number (and is very likely
given how common that is for db primary keys) which could cause
conflicts when the webhook events are sent, as the events don't have any
kind of namespacing or mapping to the original emitter, meaning if two
separate storefronts had the same approach to the external order num
then neither would be able to tell whether the event is for their order
or for another store's order.

While we only want one site to be onboarded per account there are no
system constraints in place to enforce this, so it is possible, however
unlikely, that a merchant could have multiple storefronts connected to
the same Komoju account. And given that getting the orders mixed up
could result in incorrect order handling (cancelled orders being
shipped, paid for orders being considered cancelled) and profit loss for
the merchants I think it's an important enough edge case to be
considered and designed for.

Opting to store the unique external order num in a database table,
instead of using a user supplied prefix like the WooCommerce plugin,
because changing the prefix while orders are active would break
compatibility with pending events (ie, the payment is authorised but not
captured, or refunds that may occur days/weeks later).