Add Model class for external payment table
Adding the Magento model classes so the external_payment table is
programmatically accessible in the app. Adding helper methods to the
model for creating and retrieving the object, to encapsulate the logic
around creating a unique ID to send to Komoju, and translating it back
again into the order object.