Send the external payment id instead of order id
When redirecting the user to the hosted api page the app nows creates a
entry in the komoju_external_payment table, with a unique id that is
sent as the external_order_num instead of the order id. This should
prevent issues where multiple systems are using the same Komoju account
and same approach to order numbering (ie, both are numbers that start at
1 and increment) and will allow the systems to correctly distinguish
which events belong to them, so orders aren't mistakenly moved into
incorrect states.