Add external order ID to all status messages
Prepending all order status comments with the external order Id of the
order. Because the order ID is generated by the system the Magento
admins would not be aware of what the value would be, without digging
into to the database. Because the external order num has to be unique in
Komoju this is a useful way to find the corresponding payment
information in the Komoju dashboard. To make it easier for the Magento
admins to make this connection the external order id is being added to
the order status messages.