Move search logic from Model to collection
The load() function being used to request data from the database is
deprecated, so we shouldn't use it. Switching over to using the
Collection to filter the table records according to the
external_payment_id and return the first row it finds. Since the
external_payment_id should be unique there should only be one matching
record.