Ignore webhook events that don't have records
If a webhook event is sent but doesn't have a record that matches the
event's external_order_num then log a message and ignore it. Since
multiple system can be connected to the same Komoju account it's
possible an event will be sent that doesn't correspond to any actual
orders in the system, in which case there's nothing to do.

If I don't explicity throw the error in the collection then the orderId
gets set to nothing and the order interface throws an error instead.
While this would accomplish the same effect it's a bit abstract from the
actual source.