Generate list of payments method on server side
By creating an object of active value keys to display texts for active
payment types on the server, it's much easier to provide different
languages for the options than if there were being constructed on the
server side. This also has the added benefit of being able to keep the
option values in sync with the values used to generate the Hosted Page
URL, as Komoju will only allow a strict subset of values.