Rename checkout JS and HTML files
Referring to the files as cc (Credit Card) doesn't make much sense, as
the same JS and HTML files are using for all payment types. Originally,
I thought that I'd need different templates for each payment type, but
since the billing details are a part of magento core the only info
specific to Komoju is the payment method, which is just a drop down with
a variable amount of options in it, depending on what is enabled.