Keep full page loading spinner while redirecting
Previously the full page loading spinner would disappear after Magento
had successfully save the order. However, this was leaving the page
interactable and with no loading indication when the Komoju endpoint
constructed the Hosted Page API and redirected the user to it. This
could've lead to some confusing situations where it appeared as if
nothing was happening, or could have lead the user to try submit twice,
accidentally placing a second order.

Updated the code to inject Magento core's full page loading spinner into
the Komoju form JS, so the loading spinner can be presented while the
user is waiting to be redirected.