Use base currency code for hosted page redirect
Previously the order currency code was being used to construct the
Hosted Page URL, however, that is not correct as the currency used to
charge the customer is the base currency for the store. The currency for
the order is the "display" currency. Since the store base currency is
the currency the customer should be charged in, it's being used to build
the URL.