Add event to disable the payment method
Since Komoju only has a small set of supported currencies we don't want
to allow people to try and purchase in an unsupported currency (since
that will just result in an error), so the payment method will be
disabled if the store currency is unsupported. This is being done using
Magento's internal event that determines whether to display a payment
method or not.

The event is fired on each request for the checkout page despite the
base currency being a store wide setting. I'd rather disable Komoju from
being configured in the admin settings but as far as I know that's not
possible.