Add note on valid currency types on settings
Adding comment below the active select field to let the user know that
the plugin will only work with selected currencies. I'd rather generate
the comment based on the consts in the ConfigProvider, but as far as I'm
aware that's not possible with the system.xml file, so I've settled with
a comment letting future devs know they need to update the comment when
the list of supported currencies change.