Mark order complete if refunded
If the webhook emits an event for a refunded payment, add a note to the
order and mark it complete. Marking it complete as opposed to closed as
a complete order can still create a credit memo, which is used to create
a refund in the Magento system. Since this event is only emitted when a
refunded has already been issued we don't need the credit memo to create
a refund, but allowing the admins to create one for completeness is
still a good idea.

In the future the processing the refunded event could include a step to
create a credit memo programmatically.