Disable can_void functionality
The can_void functionality allows users to cancel a processing payment.
This can be done via the Komoju dashboard, but currently not through the
plugin, so we shouldn't indicate it's possible.

I'm explicitly marking the can_void and can_refund abilities as false to
be clear to future devs that they are options and aren't supported. This
also makes it easier to know how to enable them in the future if they
become supported, instead of having to sift through the Magento
documentation to figure out the option name.