Update order total paid when payment is captured
When the payment is captured updating the order's total_paid field to
add the amount in the webhook event to the current total paid. I think
the addition step is unnecessary as Komoju only supports single
payments, but the additional code to support it is not much and it will
act as a good future proof if multiple payments is supported later.