Add new table to track processed refunds
Because the payment.refund.created event includes a list of all refund
events for the payment it's not possible to easily determine which
refund caused the event to be sent. Therefore, we need to store the
refund events in a table, so we can determine which ones have already
been processed and ignore them.