Support payment.refund.created events
The payment.refunded events are only emitting when a payment is fully
refunded. However, if a payment is partially refunded (eg, someone
bought multiple goods and decided to return one of them) the
payment.refunded event is not emitted. To support partial refunds the
app needs to process payment.refund.created events, which get emitted
every time a refund is created.

Unfortunately the payment.refund.created events contain all the refunds
for the payment, and don't specify which one is the cause of the event,
so to ensure that the same refund is not being processed multiple times,
we're tracking which refunds we've seen in the database.

For each refund generated we're creating a credit memo, which is
Magento's representation of a refund and is used by the system to track
refilling inventory, etc.