Replace custom statuses
If we want to be able to display a custom status we need to add a setup
hook to module to add them to Magento. They're not really that important
to require that amount of effort, so just setting the statuses to the
Magento defaults, so they don't display blank fields.