Translate webhook generate comments
All the comments that are created due to webhook events now have
Japanese language equivalents. Unfortunately my beginner Japanese is
_nowhere_ near good enough to translate them by hand, so I used Google,
which most certainly did not translate them fluently.

I needed to update the way the comment strings were created as the i18n
dictionary generator is specifically looking for phrases, and the
existing process of using . to combine strings and variables was not
being picked up by the command.