Fix schema
The table field for the foreign key was incorrectly referring a column
name instead of the table. This doesn't seem to effect the table setup
as it was correctly being applied to the table, but I wanted to update
the schema to reflect how the table would be generated.

Also removed the index on the sales_order_id as we aren't searching the
table with that field, so there's no reason to incur the cost of an
index at this point.