Add magento-coding-standard as dependency
Adding the magento coding standard repo as a dependency so we can use
phpcs to violations of the magento code standard. Because Magento check
this during review, it's important that we have a way of checking and
fixing this before uploading, so we don't have to wait for them to
reject the plugin first.