Add step in install eslint in the container
During setup globally install ESLint into the container, so linting can
be performed inside the docker system. I'm opting to just hardcode the
specifc install instruction and package version, instead of using
something like NPM or Yarn, because Magento uses RequireJS to handle
third party JS dependencies, and I don't want future developers to be
confused as to why packages installed into a package.json are not
accessible inside the Magento plugin itself.

If we get to the point that we have multiple JS dependencies we rely for
development, then we shoud reassess this method, but I think it's fine
just for ESLint.