Update Dockerfile to improve consistency
Copying the Dockerfile from the docker-magento repo instead of using as
a base, as we need to make changes to what exists at the /var/www/html
directory, and that has to happen before the volume declaration, to
ensure the data gets correctly copied across. Otherwise, any changes we
make inside the build process get ignored as Docker can't make changes
to a volume during build.

Instead of downloading the version of Magento externally and then
copying to the container, we're now downloading it as part of the build
process. This means we don't have additional copies of magento living in
the host system, taking up space. It also means we don't have to extract
it into the project, adding a bunch of additional files we don't need to
develop a Magento extension.

By pushing a lot of this work into the Docker build process I'm hoping to
ensure build consistency across multiple computers, as the previous
process was not working, even between 2 systems with the same OS and
shells.