Update scripts to get PHP linting working
With the change to the docker dev environment the PHP linting got broken
somehow. To fix this we're now using to the direct vendor executables,
rather than the vendor/bin symlinks since CodeSniffer uses some kind of
code reflection that's stopped working with the symlinks. As a
precaution the setup process has already been updated to set the
executable bit on all vendor/bin files, since with that permission
they're pretty useless.