Add Marketing directory to the docs
When submitting the extension to Magento for listing on their
marketplace, we need to pass a marketing review as well as a technical
review. To ensure that anyone working on this project in the future has
all the assets they need all marketing material will be included in a
docs/ folder in the repository.