Update default store name
Changing the default store name to be more descriptive of what
payments are available with KOMOJU, since it's unlikely people will
actually know what payment options are available by the company name
alone.

It doesn't appear that the default values for fields can be translated
into multiple (as far as I can tell) so going to default with Japanese
since that is the main language we think people will use with the
plugin. If it isn't then store owners can just change the default value
in the config settings.