Replace / with ・
Turns out Magento cannot process XML with a / in the value, despite it
being valid XML syntax. This resulted in the default value not being
committed to the database, so it would not show anything unless the
store owners manually set the value in the config.

Replacing / with ・ as it's a pretty common separator in Japan, and
Magento recognises the character correctly.