Add a comment to Webhook field with the URL
There's currently no indication to the store admins what the webhook URL
would be, which is required to create a webhook in the Komoju dashboard,
so I'm adding a comment underneath the webhook secret token config field
with the path. I'd love to be able to display the full path of the
webhook, but the config is in XML, and as far as I'm aware, we can't
programmatically insert data into it.