Add extension icon
Using the KOMOJU logo text as the extension icon as that seems like the
most appropriate asset to use that I could find. Magento requires the
icon to be 240x240px so rather than horribly distorting the text I kept
it at 240px width and filled in the rest with a white background.