Update linting to indicate no problems found
Currently when the linter does not find any errors it just exits with no
information. This is unclear whether there were no issues or if the
script just did not run as intended, so the script now outputs messages when
either the PHP or JS linting finishes without finding any problems.