Add section on how to create the zip file
For Magento to consider the zip file to be valid, it needs to start from
the root of the plugin (the folder with the composer.json file in it).
Since this is not obvious for non-Magento devs I wanted to add a section
explicitly detailing how to create the zip when we need to upload a new
version.