Update detailed description
We failed the marketing review because of the detailed description
formatting, so I've re-reviewed the guidelines and made some changes.
Since the feedback didn't include any specific information as to why it
failed I don't know if this will pass the review, so I guess we'll keep
updating it until it does.